requirements.txt
torch>=2.1.0
transformers>=4.40.0
datasets>=2.14.0
accelerate>=0.25.0
pyyaml>=6.0
wandb>=0.16.0
tqdm>=4.65.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
lm-eval>=0.4.0
